+++ lab3/neural_network.h
@@ -11,22 +11,22 @@ using namespace std;
 class neural_network;
 
 typedef struct {
-	vector<float> input;
-	vector<float> target;
+	vector<double> input;
+	vector<double> target;
 } training_data;
 
 class neural_network{
 private:
 	vector<layer * > layers;
 	int layer_count;	//number of layers in neural network
-	float tolerance;	//% error which can be accepted
-	float Threshold;
+	double tolerance;	//% error which can be accepted
+	double Threshold;
 	void create_edges();
-	//vector<float> get_output();
+	//vector<double> get_output();
 //methods
 public:
 	neural_network();
-	neural_network(int, vector<int>, float thresh=0.1);
+	neural_network(int, vector<int>, double thresh=0.1);
 	~neural_network();
 
 	void fwd_propogate();
@@ -35,9 +35,9 @@ public:
 	void print_network();
 	void weight_update();
 	int training_step(vector<training_data>);
-	void set_input(vector<float>);
-	float calculate_err(vector<float>);			//returns the diatance between output layer vector and vector
-	vector<float> calculate_output(vector<float>);
+	void set_input(vector<double>);
+	double calculate_err(vector<double>);			//returns the diatance between output layer vector and vector
+	vector<double> calculate_output(vector<double>);
 };
